+++ src/main/java/com/example/runshop/model/dto/product/UpdateProductRequest.java
@@ -5,10 +5,12 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @EqualsAndHashCode(callSuper = true)  // 부모 클래스의 equals, hashCode 메서드를 포함하도록 설정
 public class UpdateProductRequest extends BaseProductDTO{
-    public UpdateProductRequest(String name, String description, int price, Category category, String brand) {
+    public UpdateProductRequest(String name, String description, BigDecimal price, Category category, String brand) {
         super.setName(name);
         super.setDescription(description);
         super.setPrice(price);
